fix(kube): make state miss not a error
This is a partial shot in the dark currently, but this appears to be
what's causing the deployment issues. I've been able to reproduce this
issue quite frequently in multiple different environments and across
helm versions, so, from the looks of this this should resolve it. Open
to any feedback on a better way of handling this. Ideally since this
appears to be just checking the internal state and seeing if it exists,
we should be able to ignore that and instead make it a warning. I'll
adjust the code as needed based on usage tests.

Fixes #1193